C6 (#113): AnimatedStatCard unique remplace les 3 copies stats-row
Le StatCard motion copié dans users/exams/questions-stats-row (~130 lignes × 3)
devient components/admin/animated-stat-card.tsx (5 couleurs, trend optionnel,
skeleton exporté). Les rows ne gardent que leur config. stat-card.tsx (famille
shadcn testée) et les variantes ad hoc (exam-stat-card, transaction-stats)
restent volontairement distincts.
